feat(backend): migrate from SQLite to PostgreSQL
- sqlx feature sqlite -> postgres; PgPool throughout
- migrations ported (identity columns, BIGINT, DOUBLE PRECISION,
  DROP CONSTRAINT instead of table rebuilds)
- placeholders converted to $N; pg_params() renumbers dynamically
  assembled SQL
- dialect fixes: RETURNING id, ON CONFLICT upserts, ::timestamptz
  comparisons, EXTRACT(EPOCH ...) durations, GREATEST, ILIKE searches,
  jsonb_build_object / jsonb || merge, aggregate casts (::bigint/::float8),
  boolean CASE binds, typed NULL binds
- tests create one throwaway PG database each (TEST_DATABASE_ADMIN_URL,
  default postgres://localhost/postgres); stale mcss_test_* swept
- CI runs against a postgres:18 service container
- docker-compose gains a postgres service; Dockerfile/.env/README updated

All 545 tests pass; clippy-clean.

+++ backend/migrations/0003_activity.sql
@@ -2,7 +2,7 @@
 -- Restore bookkeeping columns: the guarded one-time transition that makes a
 -- restore exactly-once observable as not-restored / in-progress / restored.
 ALTER TABLE activity_logs ADD COLUMN restore_state TEXT;          -- NULL | 'in_progress' | 'restored'
-ALTER TABLE activity_logs ADD COLUMN restored_by_log_id INTEGER;  -- restore audit entry id (link)
+ALTER TABLE activity_logs ADD COLUMN restored_by_log_id BIGINT;  -- restore audit entry id (link)
 ALTER TABLE activity_logs ADD COLUMN restored_at TEXT;
 
 CREATE INDEX idx_activity_logs_created_at ON activity_logs(created_at);
